python_face/requirements.txt
opencv-contrib-python
pillow
pymongo
numpy
Flask
flask-cors